requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
flask>=3.0.0
streamlit>=1.28.0
redis>=5.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
